refactor: 💡 moving constants to constants folder
✅ Closes: DA-14
