+++ src/components/pages/Character/index.tsx
@@ -10,10 +10,11 @@ import {
 import { characterService } from "@/connection";
 import { useCharacter } from "@/hooks/useCharacter";
 import { showToast } from "@/providers";
-import { Character, User, defaultCharacter } from "@/types";
+import { Character, User } from "@/types";
 import { useEffect, useState } from "react";
 import { useReadLocalStorage } from "usehooks-ts";
 import { Container } from "./styles";
+import { defaultCharacter } from "@/constants";
 
 export function Character() {
   const [character, setCharacter] = useState<Character>(defaultCharacter);
