+++ src/types/character.ts
@@ -80,66 +80,3 @@ export type Skill = {
   name: string;
   attribute: string;
 };
-
-interface AttributeMapper {
-  [key: string]: keyof Character["attributes"];
-}
-
-export const attributeMap: AttributeMapper = {
-  DES: "dexterity",
-  FOR: "strength",
-  CON: "constitution",
-  INT: "intelligence",
-  SAB: "wisdom",
-  CAR: "charisma",
-};
-
-export const defaultSpell = {
-  name: "",
-  level: 0,
-  description: "",
-  upgrade: "",
-  school: "",
-  castingTime: "",
-  range: "",
-  duration: "",
-  ritual: false,
-  concentration: false,
-  classes: "",
-  damage: {
-    type: "",
-    characterLevel: {},
-    slotLevel: {},
-  },
-};
-
-export const defaultCharacter = {
-  _id: "",
-  name: "",
-  attributes: {
-    strength: 0,
-    dexterity: 0,
-    constitution: 0,
-    intelligence: 0,
-    wisdom: 0,
-    charisma: 0,
-  },
-  armorClass: 0,
-  speed: 0,
-  level: 1,
-  experience: 0,
-  hitPoints: 0,
-  maxHitPoints: 0,
-  status: [],
-  languages: [],
-  spells: [],
-  userId: "",
-  class: "",
-  race: "",
-  skills: [],
-  equipments: [],
-  gold: 0,
-  quests: [],
-  inspiration: 0,
-  features: [],
-};
